+++ README.md
@@ -1,0 +1,5 @@
+This repository contains implementation of practice problems we discuss in weekly live classes.
+
+| Week |                         Video                          |          Problem 1           |            Problem 2             |
+| :--: | :----------------------------------------------------: | :--------------------------: | :------------------------------: |
+|  1   | [Youtube](https://www.youtube.com/watch?v=Nmn6gnsfIpM) | [Divisor](./Week-1/divisor/) | [Increment](./Week-1/increment/) |
