--- Week-1/divisor.cpp
@@ -1,59 +1,0 @@
-/***
-Problem statement:
-You are given integers N and M.
-Consider a sequence of length N consisting of positive integers such that a1​+a2​+...+aN​ = M. 
-Find the maximum possible value of the greatest common divisor of a1​,a2​,...,aN​.
-
-Constraints:
-    All values in input are integers.
-	1≤N≤10^5
-    N≤M≤10^9
-
-Input:
-Input is given from standard input in the following format:
-N M
-
-Output:
-Print the maximum possible value of the greatest common divisor of a sequence a1​,a2​,...,aN​ that satisfies the condition.
-
-Examples:
-	Sample Input 1
-	3 14
-
-	Sample Output 1
-	2
-
-	Explanation
-	Consider the sequence (a1​,a2​,a3​)=(2,4,8). Their greatest common divisor is 2, and this is the maximum value. 
-	In particular, there is no sequence involving numbers larger than two for which the sum is 14 and the GCD is larger than two.
-****/
-
-
-#include <bits/stdc++.h>
-using namespace std;
-
-vector<int> getDivisors(int M) {
-    vector<int> divisors;
-    
-	/* Find the divisors of M and store in vector*/
-    for(int i=1;i<M;i++) {
-        if (M%i == 0)
-            divisors.push_back(i);
-    }
-    return divisors;
-}
-
-int main() {
-	int N, M;
-	cin >> N >> M;
-	vector<int> divisors = getDivisors(M);
-	int ans = 1;
-	/* Find the largest divisor D which satisfies the condition M >= N*D */
-	for(int i=0; i < divisors.size(); i++) {
-	    if (N <= M / divisors[i]) {
-	        ans = divisors[i];
-	    }
-	}
-	cout << ans <<"\n";
-	return 0;
-}
